Refactor sources more
